Rename Aggregate to EventSource
