+++ src/Application/Account/Handler/WithdrawMoneyCommandHandler.php
@@ -5,14 +5,14 @@ namespace App\Application\Account\Handler;
 use App\Application\Account\Event\AccountUpdated;
 use App\Application\Account\Command\WithdrawMoneyCommand;
 use App\Domain\Account\AccountAggregate;
-use App\Infrastructure\Aggregator\AggregateRepositoryInterface;
+use App\Infrastructure\EventSource\EventSourceRepositoryInterface;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
 
 final class WithdrawMoneyCommandHandler implements MessageHandlerInterface
 {
     /**
-     * @var AggregateRepositoryInterface
+     * @var EventSourceRepositoryInterface
      */
     private $aggregateRepository;
 
@@ -21,7 +21,7 @@ final class WithdrawMoneyCommandHandler implements MessageHandlerInterface
      */
     private $bus;
 
-    public function __construct(AggregateRepositoryInterface $aggregateRepository, MessageBusInterface $bus)
+    public function __construct(EventSourceRepositoryInterface $aggregateRepository, MessageBusInterface $bus)
     {
         $this->aggregateRepository = $aggregateRepository;
         $this->bus = $bus;
