+++ src/Domain/Account/AccountAggregate.php
@@ -6,8 +6,8 @@ use App\Domain\Account\Event\AccountClosed;
 use App\Domain\Account\Event\AccountCredited;
 use App\Domain\Account\Event\AccountDebited;
 use App\Domain\Account\Event\AccountOpened;
-use App\Infrastructure\Aggregator\AggregateRoot;
-use App\Infrastructure\Aggregator\Changed;
+use App\Infrastructure\EventSource\AggregateRoot;
+use App\Infrastructure\EventSource\Changed;
 use App\Infrastructure\Money\Currency;
 use App\Infrastructure\Money\Money;
 use App\Infrastructure\UUID;
