+++ src/Infrastructure/EventSource/Metadata.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Infrastructure\Aggregator;
+namespace App\Infrastructure\EventSource;
 
 final class Metadata
 {
