+++ src/Infrastructure/Resources/config/doctrine/infrastructure/EventSource.Changed.orm.xml
@@ -4,7 +4,7 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                           https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="App\Infrastructure\Aggregator\Changed" table="event_source" repository-class="App\Infrastructure\Aggregator\AggregateRepository">
+    <entity name="App\Infrastructure\EventSource\Changed" table="event_source" repository-class="App\Infrastructure\EventSource\EventSourceRepository">
 
         <unique-constraints>
             <unique-constraint columns="event_id" name="event_idx" />
@@ -24,7 +24,7 @@
         <field name="createdAt" column="created_at" type="datetime" />
 
         <entity-listeners>
-            <entity-listener class="App\Infrastructure\Aggregator\AggregateRootListener">
+            <entity-listener class="App\Infrastructure\EventSource\EventSourceListener">
                 <lifecycle-callback type="postPersist"   method="postPersistHandler"/>
             </entity-listener>
         </entity-listeners>
