+++ tests/Application/Account/Handler/CloseAccountHandlerTest.php
@@ -5,8 +5,8 @@ namespace App\Tests\Application\Account\Handler;
 use App\Application\Account\Command\CloseAccount;
 use App\Application\Account\Handler\CloseAccountHandler;
 use App\Domain\Account\AccountAggregate;
-use App\Infrastructure\Aggregator\AggregateRepositoryInterface;
-use App\Infrastructure\Aggregator\AggregateRoot;
+use App\Infrastructure\EventSource\EventSourceRepositoryInterface;
+use App\Infrastructure\EventSource\AggregateRoot;
 use PHPUnit\Framework\TestCase;
 use App\Infrastructure\UUID;
 use Prophecy\Argument;
@@ -22,7 +22,7 @@ final class CloseAccountHandlerTest extends TestCase
         $accountAggregate->close()->shouldBeCalled();
         $accountAggregate->getId()->willReturn($id)->shouldBeCalled();
 
-        $repo = $this->prophesize(AggregateRepositoryInterface::class);
+        $repo = $this->prophesize(EventSourceRepositoryInterface::class);
         $repo->load($id, AccountAggregate::class)->willReturn($accountAggregate->reveal())->shouldBeCalled();
         $repo->store(
             Argument::that(
