+++ tests/Domain/Account/AccountAggregateTest.php
@@ -5,8 +5,8 @@ namespace App\Tests\Domain\Account;
 use App\Domain\Account\AccountAggregate;
 use App\Domain\Account\Event\AccountCredited;
 use App\Domain\Account\Event\AccountOpened;
-use App\Infrastructure\Aggregator\Changed;
-use App\Infrastructure\Aggregator\Metadata;
+use App\Infrastructure\EventSource\Changed;
+use App\Infrastructure\EventSource\Metadata;
 use App\Infrastructure\Money\Currency;
 use App\Infrastructure\Money\Money;
 use PHPUnit\Framework\TestCase;
